Refactor HeapBalancer into LeastLoadBalancer.
It's actually not doable to have a HeapBalancer with the current design.
The availability is changing independently from the "LoadBalancer load", keeping
an the ServiceFactories ordered imply sorting them at every access.

Then the creation of the LeastLoadedBalancer.
This should behave ok when the number of ServiceFactories is small, but I will
introduce later another LoadBalancer based on "The Power of Two Choices".